Reafctored pseudoContentRegex so that IP Addresses remain their frmat an get pseudonymized with CryptoPAn
